ChatBot/requirements.txt
pyrebase==3.0.27
pycryptodome==3.4.3
Flask==3.0.2
boto3==1.34.55
python-dotenv==1.0.1